final bug fixes on map, and refactoring, ready for merge
